most of the way through v0.2.4; diffloop integration and significant reworking of README

+++ hichipper/hichipper.sh
@@ -15,30 +15,44 @@ MAX_DIST=${11}
 MACS2_STRING=${12}
 
 # Make useful shortcuts
-LOG_FILE="${OUT_NAME}/hichipper.log"
+LOG_FILE="${OUT_NAME}/${OUT_NAME}.hichipper.log"
 
 echo "`date`: Processing ${SAMPLE}" | tee -a $LOG_FILE
 
 # Process .bam into bedpe
+if [ ! -f "${BAM_ONE}" ] ; then
+    echo "File ${BAM_ONE} is not in ${OUT_NAME}, aborting." | tee -a $LOG_FILE
+    exit
+fi
+
+if [ ! -f "${BAM_TWO}" ] ; then
+    echo "File ${BAM_TWO} is not in ${OUT_NAME}, aborting." | tee -a $LOG_FILE
+    exit
+fi
+
+# Do stuff with samtools
 samtools view -F2304 "${BAM_ONE}" | awk -v READ_LEN="$READ_LEN" -v MIN_QUAL="$MIN_QUAL" '{if ($5>=MIN_QUAL) print $3,$4,$4+READ_LEN,$1; else print "*","*","*",$1}' OFS='\t' > "${OUT_NAME}/${SAMPLE}_pos_r1.bed.tmp"
 samtools view -F2304 "${BAM_TWO}" | awk -v READ_LEN="$READ_LEN" -v MIN_QUAL="$MIN_QUAL" '{if ($5>=MIN_QUAL) print $3,$4,$4+READ_LEN,$1; else print "*","*","*",$1}' OFS='\t' > "${OUT_NAME}/${SAMPLE}_pos_r2.bed.tmp"
+if [ ! -f "${OUT_NAME}/${SAMPLE}_pos_r2.bed.tmp" ] ; then
+    echo "File specified doesn't seem like a .bam and/or samtools is not accessible, aborting." | tee -a $LOG_FILE
+    exit
+fi
+
+# The example files that I generated aren't actually paired, so I'm putting this condition as a warning
 Total_PETs1=`samtools flagstat "${WK_DIR}/${BAM_ONE}" | head -1 | awk '{print $1}'`
 Total_PETs2=`samtools flagstat "${WK_DIR}/${BAM_TWO}" | head -1 | awk '{print $1}'`
-
-# The example files that I generated aren't actually paired, so I'm putting this condition in there
 if [ "$Total_PETs1" -lt "$Total_PETs2" ]; then
-	echo "WARNING-- .bam 1 has more reads than .bam 2. Check to see if they actually match!" | tee -a $LOG_FILE
+	echo "`date`: WARNING-- .bam 1 has more reads than .bam 2. Check to see if they actually match!" | tee -a $LOG_FILE
 	head "-${Total_PETs1}" "${OUT_NAME}/${SAMPLE}_pos_r2.bed.tmp" > temppets
 	mv temppets "${OUT_NAME}/${SAMPLE}_pos_r2.bed.tmp"
 elif [ "$Total_PETs2" -lt "$Total_PETs1" ]; then
-	echo "WARNING-- .bam 2 has more reads than .bam 1. Check to see if they actually match!" | tee -a $LOG_FILE
+	echo "`date`: WARNING-- .bam 2 has more reads than .bam 1. Check to see if they actually match!" | tee -a $LOG_FILE
 	head "-${Total_PETs2}" "${OUT_NAME}/${SAMPLE}_pos_r1.bed.tmp" > temppets
 	mv temppets "${OUT_NAME}/${SAMPLE}_pos_r1.bed.tmp"
 fi
 
+# Pair anyways; keep reads that are both mapped
 paste "${OUT_NAME}/${SAMPLE}_pos_r1.bed.tmp" "${OUT_NAME}/${SAMPLE}_pos_r2.bed.tmp" | cut -f 1-3,5-8 > "${OUT_NAME}/${SAMPLE}_rawinteractions.tmp" 
-
-# Keep only interactions where both reads are mapped
 awk '{if ($1 != "*" && $4 != "*") print}' "${OUT_NAME}/${SAMPLE}_rawinteractions.tmp" > "${OUT_NAME}/${SAMPLE}_interactions.unsorted.bedpe.tmp"
 
 # Swap anchors if necessary
@@ -53,16 +67,21 @@ Mapped_unique_PETs_q30=`wc -l "${OUT_NAME}/${SAMPLE}_interactions.dedup.bedpe.tm
 Mapped_unique_intrachromosomal_q30=`awk '$1 == $4 {print $0}' "${OUT_NAME}/${SAMPLE}_interactions.dedup.bedpe.tmp" | wc -l | awk '{print $1}'`
 Mapped_unique_intrachromosomal_q30_5kb=`awk -v MIN_DIST="$MIN_DIST" '$1 == $4 && (($5+$6) - (($2+$3)/2)>=MIN_DIST) {print $0}' "${OUT_NAME}/${SAMPLE}_interactions.dedup.bedpe.tmp" | wc -l | awk '{print $1}'`
 
-echo "Mapped_unique_PETs_q30=${Mapped_unique_PETs_q30}" 
-echo "Mapped_unique_intrachromosomal_q30=${Mapped_unique_intrachromosomal_q30}"
-echo "Mapped_unique_intrachromosomal_q30_5kb=${Mapped_unique_intrachromosomal_q30_5kb}" 
-
+echo "`date`: Mapped_unique_PETs_q30=${Mapped_unique_PETs_q30}" | tee -a $LOG_FILE
+echo "`date`: Mapped_unique_intrachromosomal_q30=${Mapped_unique_intrachromosomal_q30}"| tee -a $LOG_FILE
+echo "`date`: Mapped_unique_intrachromosomal_q30_5kb=${Mapped_unique_intrachromosomal_q30_5kb}" | tee -a $LOG_FILE
 
 # Call peaks
 cut -f 1-3,7 "${OUT_NAME}/${SAMPLE}_interactions.dedup.bedpe.tmp" >  "${OUT_NAME}/${SAMPLE}_left.dedup.bed.tmp"
 cut -f 4-6,7 "${OUT_NAME}/${SAMPLE}_interactions.dedup.bedpe.tmp" >  "${OUT_NAME}/${SAMPLE}_right.dedup.bed.tmp"
 cat "${OUT_NAME}/${SAMPLE}_left.dedup.bed.tmp" "${OUT_NAME}/${SAMPLE}_right.dedup.bed.tmp" >> "${OUT_NAME}/${SAMPLE}_reads.bed.tmp"	
-macs2 callpeak -t "${OUT_NAME}/${SAMPLE}_reads.bed.tmp" -g hs -f BED -n "${OUT_NAME}/${SAMPLE}_temporary" $MACS2_STRING
+macs2 callpeak -t "${OUT_NAME}/${SAMPLE}_reads.bed.tmp" -g hs -f BED -n "${OUT_NAME}/${SAMPLE}_temporary" $MACS2_STRING | tee -a $LOG_FILE
+
+# Check to see if MACS2 worked
+if [ ! -f "${OUT_NAME}/${SAMPLE}_temporary_peaks.narrowPeak" ] ; then
+    echo "MACS2 execution was not successful--modify command if possible" | tee -a $LOG_FILE
+    exit
+fi
 
 # Pad peaks
 awk -v PEAK_PAD="$PEAK_PAD" '{$2-=PEAK_PAD; $3+=PEAK_PAD}1' OFS="\t" "${OUT_NAME}/${SAMPLE}_temporary_peaks.narrowPeak" > "${OUT_NAME}/${SAMPLE}_temporary_peaks.narrowPeak.padded"
@@ -70,6 +89,7 @@ echo "`date`: Padded peaks by ${PEAK_PAD}bp" | tee -a $LOG_FILE
 awk '$2 > 0 && $5 > 0 {print $0}' "${OUT_NAME}/${SAMPLE}_temporary_peaks.narrowPeak.padded" > "${OUT_NAME}/${SAMPLE}_temporary_peaks.narrowPeak.padded.clean"
 
 # Merge gaps
+echo "`date`: Intersecting PETs with anchors" | tee -a $LOG_FILE
 bedtools merge -d $MERGE_GAP -i "${OUT_NAME}/${SAMPLE}_temporary_peaks.narrowPeak.padded.clean" > "${OUT_NAME}/${SAMPLE}_temporary_peaks.merged.bed.tmp"
 bedtools intersect -loj -a "${OUT_NAME}/${SAMPLE}_left.dedup.bed.tmp" -b "${OUT_NAME}/${SAMPLE}_temporary_peaks.merged.bed.tmp" | awk '{print $5,$6,$7,$4}' OFS='\t' > "${OUT_NAME}/${SAMPLE}_anchor1.bed.tmp"
 bedtools intersect -loj -a "${OUT_NAME}/${SAMPLE}_right.dedup.bed.tmp" -b "${OUT_NAME}/${SAMPLE}_temporary_peaks.merged.bed.tmp" | awk '{print $5,$6,$7,$4}' OFS='\t' > "${OUT_NAME}/${SAMPLE}_anchor2.bed.tmp"
@@ -78,10 +98,18 @@ bedtools intersect -loj -a "${OUT_NAME}/${SAMPLE}_right.dedup.bed.tmp" -b "${OUT
 paste "${OUT_NAME}/${SAMPLE}_anchor1.bed.tmp" "${OUT_NAME}/${SAMPLE}_anchor2.bed.tmp" | cut -f 1-3,5-8 > "${OUT_NAME}/${SAMPLE}_anchor.interactions.tmp"
 awk '{if ($1 != "." && $4 != ".") print}' "${OUT_NAME}/${SAMPLE}_anchor.interactions.tmp" > "${OUT_NAME}/${SAMPLE}_anchor.interactions.bedpe.tmp"
 cut -f1-6  "${OUT_NAME}/${SAMPLE}_anchor.interactions.bedpe.tmp" | sort | uniq -c | awk '{print $2,$3,$4,$5,$6,$7,".",$1}' >  "${OUT_NAME}/${SAMPLE}.loop_counts.bedpe.tmp"
-Mapped_unique_intrachromosomal_q30_5kb=`awk -v MIN_DIST="$MIN_DIST" '$1 == $4 && (($5+$6)/2 - (($2+$3)/2)>=MIN_DIST) {print $0}' "${OUT_NAME}/${SAMPLE}_interactions.dedup.bedpe.tmp" | wc -l | awk '{print $1}'`
+Mapped_unique_intrachromosomal_q30_5kb=`awk -v DIST=5000 '$1 == $4 && (($5+$6)/2 - (($2+$3)/2)>=DIST) {print $0}' "${OUT_NAME}/${SAMPLE}_interactions.dedup.bedpe.tmp" | wc -l | awk '{print $1}'`
+Anchor_mapped_PETs_5kb=`awk -v DIST=5000 '$1 == $4 && (($5+$6)/2 - (($2+$3)/2)>=DIST) {sum += $8} END {print sum}' "${OUT_NAME}/${SAMPLE}.loop_counts.bedpe.tmp"`
 
+echo "`date`: Mapped_unique_intrachromosomal_q30_5kb=${Mapped_unique_intrachromosomal_q30_5kb}"| tee -a $LOG_FILE
+echo "`date`: Anchor_mapped_PETs_5kb=${Anchor_mapped_PETs_5kb}" | tee -a $LOG_FILE
+
+# Produce final output
 awk '$1 != $4 {print $0}' "${OUT_NAME}/${SAMPLE}.loop_counts.bedpe.tmp" > "${OUT_NAME}/${SAMPLE}.inter.loop_counts.bedpe"
-awk '$1 == $4 && $2 != $5 {print $0}' "${OUT_NAME}/${SAMPLE}.loop_counts.bedpe.tmp" > "${OUT_NAME}/${SAMPLE}.intra.loop_counts.bedpe"
+awk '$1 == $4 {print $0}' "${OUT_NAME}/${SAMPLE}.loop_counts.bedpe.tmp" > "${OUT_NAME}/${SAMPLE}.intra.loop_counts.bedpe"
+awk -v MIN_DIST="$MIN_DIST" -v MAX_DIST="$MAX_DIST" '$1 == $4 && $2 != $5 && (($5+$6)/2 - ($2+$3)/2)>=MIN_DIST && (($5+$6)/2 - ($2+$3)/2)<=MAX_DIST {print $0}' "${OUT_NAME}/${SAMPLE}.loop_counts.bedpe.tmp" > "${OUT_NAME}/${SAMPLE}.filt.intra.loop_counts.bedpe"
+Loop_PETs=`awk '{sum += $8} END {print sum}' "${OUT_NAME}/${SAMPLE}.filt.intra.loop_counts.bedpe"`
+echo "`date`: Loop_PETs=${Loop_PETs}" | tee -a $LOG_FILE
 
 # Write out summary statistics
 echo "Total_PETs=${Total_PETs1}" > "${OUT_NAME}/${SAMPLE}.stat"
@@ -89,3 +117,6 @@ echo "Mapped_PETs_q30=${Mapped_PETs_q30}" >> "${OUT_NAME}/${SAMPLE}.stat"
 echo "Mapped_unique_PETs_q30=${Mapped_unique_PETs_q30}" >> "${OUT_NAME}/${SAMPLE}.stat"
 echo "Mapped_unique_intrachromosomal_q30=${Mapped_unique_intrachromosomal_q30}" >> "${OUT_NAME}/${SAMPLE}.stat"
 echo "Mapped_unique_intrachromosomal_q30_5kb=${Mapped_unique_intrachromosomal_q30_5kb}" >> "${OUT_NAME}/${SAMPLE}.stat"
+echo "Anchor_mapped_PETs_5kb=${Anchor_mapped_PETs_5kb}" >> "${OUT_NAME}/${SAMPLE}.stat"
+echo "Loop_PETs=${Loop_PETs}" >> "${OUT_NAME}/${SAMPLE}.stat"
+
